requirements.txt
langchain>=1.2.12
langchain-chroma>=1.1.0
langchain-classic>=1.0.2
langchain-cli>=0.0.37
langchain-community>=0.4.1
langchain-core>=1.2.18
langchain_dashscope>=0.1.8
langchain-experimental>=0.4.1
langchain-openai>=1.1.11
langchain-text-splitters>=1.1.1
tavily-python
dashscope>=1.25.5
openai>=2.14.0
chromadb>=1.4.0
concurrent-log-handler>=0.9.28
SQLAlchemy>=2.0.39
python-dotenv>=1.2.1
pydantic>=2.11.7
typing-extensions>=4.12.2
minio>=7.0.0
pypdf>=3.0.0
python-docx>=0.8.11
python-pptx>=0.6.21
beautifulsoup4>=4.12.0
nbformat>=5.0.0
PyPDF2>=3.0.0
PyMuPDF>=1.23.0
chardet>=5.0.0
